refactor: stomp 리펙토링
